refactor(inventory): merge inventory model into services/inventory.rs
The inventory data types (Host, HostVars, Inventory, etc.) lived in
`src/models/inventory.rs` while the loader functions that produce
them lived in `src/services/inventory.rs`. The split was vestigial:
no other module needs the types without going through the loader.
Co-locate them in the services file and update every caller to
import from the new path.

Refs #266
