+++ src/commands/select.rs
@@ -1,6 +1,5 @@
-use crate::models::inventory::Host;
 use crate::prompt::select_item;
-use crate::services::inventory::{get_hosts, get_playbooks};
+use crate::services::inventory::{Host, get_hosts, get_playbooks};
 use clap::Subcommand;
 use eyre::Result;
 use std::path::PathBuf;
