+++ src/models.rs
@@ -1,3 +1,2 @@
-pub mod inventory;
 pub mod key_registry;
 pub mod playbook_meta;
